Changed the way that the drawing requests from inside Visvis and from the Gui toolkit result into draws. This makes the wx backend work again on Linux (but I still need to test this). Moreover it's simply cleaner.
A nice side affect is that now the fltk backend also seems stable (if I do not use any other backend in IEP). This is something I wanted for quite some time. I still should like use it for a while to see whether it really is stable though (and test on Linux)...

Also started a document with pointers how to implement a backend, and explaining how the backend interacts with visvis.