Give ThreeDCamera option of a perspective view.
The ThreeDCamera gains the view parameter view_fov.  If it is 0 (default),
an orthographic view is used.  Otherwise, a perspective view is used, with
view_fov giving the vertical field of view.  The camera is placed at a
distance from view_loc such that the plane normal to the camera direction
has the same extent in both the perspective and orthographic views.

CartesianAxis3D assumed that parallel axes were parallel in screen
coordinates, which is not true in the perspective view.  This has been
fixed for drawing the edge lines.  However, choosing which line to draw
the ticks on, the tick spacing, and the lines draw when _showBox is
False all still need to be fixed for the perspective view.