Add convenience function mesh()
The main job of mesh() is to create a Mesh with the appropriate arguments.
But it also does some input checking and can set colormaps or textures.

checkDimsOfArray (in polygonalModeling) has been modified to try coercing
its input into a numpy array.  Additionally, it will accept a 1D array if
0 is in ndims.