Axes.GetLimits returns the axes limits with the current camera.
With a 2D camera, it returns the view extents, as before.  But with a 3D
camera, the extents of the coordinate axes are returned.  Note that
these are not necessarily connected to the view.

As part of this, the calculation was moved into a new camera method, also
called GetLimits.